# Conflicts:
#	src/core/smallmap_type.hpp
#	src/lang/korean.txt
#	src/network/network_content.cpp
#	src/smallmap_gui.cpp
#	src/station_cmd.cpp

+++ src/bridge_gui.cpp
@@ -156,7 +156,7 @@ public:
 		this->bridges->NeedResort();
 		this->SortBridgeList();
 
-		this->vscroll->SetCount(bl->size());
+		this->vscroll->SetCount((uint)bl->size());
 	}
 
 	~BuildBridgeWindow()
